requirements.txt
Django==2.1.5
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
